requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.6
django-filter==25.1
django-tracking2==0.5.1
DjangoSEO==1.0
idna==3.10
kavenegar==1.1.2
python-slugify==8.0.4
requests==2.32.5
South==1.0.2
sqlparse==0.5.3
text-unidecode==1.3
urllib3==2.5.0
